# docs/swagger.yaml
basePath: /api
definitions:
  dto.Applicant:
    description: Заявитель. Поле name игнорируется при создании заявки.
    properties:
      id:
        description: уникальный идентификатор заявителя при Post запросе на создание
          заявки ОБЯЗАТЕЛЕН
        type: string
      name:
        description: Название заявителя игнорируется при создании заявки
        type: string
    type: object
  dto.AuthCredentials:
    description: Структура для авторизации пользователя. Все поля обязательны к заполнению
    properties:
      login:
        description: Логин для входа в систему, должен быть уникальным
        type: string
      password:
        description: Пароль для входа в систему
        type: string
    type: object
  dto.AuthResponse:
    description: Структура для ответа с данными пользователя
    properties:
      message:
        description: просто сообщение
        type: string
      status:
        description: обычно ok  при ошибках error
        type: string
      tokens:
        allOf:
        - $ref: '#/definitions/dto.AuthTokens'
        description: токены авторизации
      user:
        allOf:
        - $ref: '#/definitions/dto.UserProfile'
        description: данные пользователя
    type: object
  dto.AuthTokens:
    description: Структура c токенами
    properties:
      access_token:
        type: string
      refhresh_token:
        type: string
    type: object
  dto.ChangePasswordRequest:
    description: Структура для смены пароля пользователя. Все поля обязательны к заполнению
    properties:
      login:
        description: Логин для входа в систему, должен быть уникальным
        type: string
      new_password:
        description: Новый пароль для входа в систему
        type: string
      old_password:
        description: Старый пароль для входа в систему
        type: string
    type: object
  dto.ChangeProfileRequest:
    description: Структура для отправки запроса на изменение данных профиля
    properties:
      full_name:
        description: Полное имя или название организации
        type: string
      id:
        description: ID пользователя
        type: string
      login:
        description: Логин для входа в систему, должен быть уникальным
        type: string
    type: object
  dto.ChangeStatusRequest:
    properties:
      id:
        description: ID заявки
        type: string
      status:
        description: Новый статус заявки (новая, в работе, выполнена, отменена)
        type: string
    type: object
  dto.RequestFull:
    description: Полная информация о заявке на отлов бродячей собаки. Поля id, number,
      status и created_at игнорируются при создании заявки.
    properties:
      address:
        description: Адрес, где была замечена бродячая собака
        type: string
      applicant:
        allOf:
        - $ref: '#/definitions/dto.Source'
        description: '"!!!!ВАЖНО название поля json applicant и source поменяты местами!!"Источник
          информации от кого был получен запрос ID находится в справочнике'
      behavior:
        description: Поведение собаки
        type: string
      contact_person:
        description: Контактные данные
        type: string
      created_at:
        description: Дата и время создания заявки генерируется автоматически
        type: string
      dogs_count:
        description: Количество собак
        type: integer
      id:
        description: уникальный идентификатор заявки при Post запросе не указывается
        type: string
      number:
        description: уникальный номер заявки, генерируется автоматически при создании
          заявки
        type: string
      source:
        allOf:
        - $ref: '#/definitions/dto.Applicant'
        description: '"!!!!ВАЖНО название поля json applicant и source поменяты местами!!"Заявитель
          ID находится в справочнике'
      status:
        description: Статус заявки (новая, в работе, выполнена, отменена) генерируется
          после создания заявки
        type: string
      urgency:
        description: срочность
        type: string
    type: object
  dto.Response:
    description: Структура для ответа с где обычно ок и сообщние
    properties:
      message:
        description: просто сообщение
        type: string
      status:
        description: обычно ok  при ошибках error
        type: string
    type: object
  dto.ResponseUrl:
    description: Структура для ответа с где обычно ок и url
    properties:
      status:
        description: обычно ok  при ошибках error
        type: string
      url:
        description: url на скаччивание файла
        type: string
    type: object
  dto.Source:
    description: Источник информации от кого был получен запрос. Поле name игнорируется
      при создании заявки.
    properties:
      id:
        description: уникальный идентификатор источника при Post запросе на создание
          заявки ОБЯЗАТЕЛЕН
        type: string
      name:
        description: Название источника игнорируется при создании заявки
        type: string
    type: object
  dto.UserProfile:
    description: Структура для получения профиля пользователя. Все поля обязательны
      к заполнению
    properties:
      full_name:
        description: Полное имя или название организации
        type: string
      id:
        description: ID пользователя
        type: string
      login:
        description: Логин для входа в систему, должен быть уникальным
        type: string
      role:
        description: Роль пользователя, например, "admin", "user", "otdel" и т.д.
        type: string
    type: object
host: localhost:8091
info:
  contact: {}
  description: API сервиса для отлова бродячих собак.
  title: Go Dog Trapping API
  version: "1.0"
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Смена данных профлия пользователя.
      parameters:
      - description: Change Profile information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Информация успешна изменена.
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Ошибка при cмене  данных.
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Change Profile Info
      tags:
      - User
  /auth/refresh:
    post:
      description: Обновление токенов доступа и обновления.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthTokens'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.AuthTokens'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.AuthTokens'
      security:
      - BearerAuth: []
      summary: Refresh Token
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя.
      parameters:
      - description: Account credentials
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.AuthCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.AuthResponse'
      summary: Sign In
      tags:
      - Auth
  /requests:
    get:
      description: Получение всех запросов на отлов бродячих собак. Доступно только
        для районных администраторов.
      produces:
      - application/json
      responses:
        "200":
          description: Все заявки"  query param otdel_id - для получения заявок по
            отделу, доступно только для  админа
          schema:
            items:
              $ref: '#/definitions/dto.RequestFull'
            type: array
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Ошибка при получении запросов.
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get All Requests
      tags:
      - Requests
    post:
      consumes:
      - application/json
      description: Отправка запроса на отлов бродячей собаки. Поля source_id и applicant_id
        заполнять ID из справочников. Поля name в этих полях игнорируются при отправке
        запроса. Доступно для всех авторизованных пользователей.
      parameters:
      - description: Request information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RequestFull'
      produces:
      - application/json
      responses:
        "200":
          description: Запрос успешно отправлен
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Ошибка при отправке запроса
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Send Request
      tags:
      - Requests
  /requests/change-status:
    post:
      consumes:
      - application/json
      description: Изменение статуса заявки на отлов бродячей собаки. Поля id и status
        обязательны к заполнению.
      parameters:
      - description: Change Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Статус заявки успешно изменен.
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Ошибка при изменении статуса заявки.
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Change Status Request
      tags:
      - Requests
  /requests/download_act:
    get:
      description: Получение URL для скачивания файла акта выполненого отлова в формате
        .docx. Требуется параметры number и year в query, номер заявки и год. Доступно
        для всех авторизованных пользователей.
      parameters:
      - description: Request Number
        in: query
        name: number
        required: true
        type: string
      - description: Request Year
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseUrl'
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Ошибка при получении запросов.
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Download Act File URL
      tags:
      - Requests
  /requests/download_request:
    get:
      description: Получение URL для скачивания файла заявки в формате .docx. Требуется
        параметр number в query, номер заявки. Доступно для всех авторизованных пользователей.
      parameters:
      - description: Request Number
        in: query
        name: number
        required: true
        type: string
      - description: Request Year
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseUrl'
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Ошибка при получении запросов.
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Download Request File URL
      tags:
      - Requests
  /requests/upload_act:
    post:
      consumes:
      - multipart/form-data
      description: Загрузка акта выполненого отлова любой формат файла, доступно админу
        и подрядчику.
      parameters:
      - description: Request Number
        in: formData
        name: number
        required: true
        type: string
      - description: Request ID
        in: formData
        name: id
        required: true
        type: string
      - description: Request Status
        in: formData
        name: status
        required: true
        type: string
      - description: Act File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Файл успешно загружен.
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Ошибка в данных запроса.
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Ошибка при загрузке акта.
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Upload Act File
      tags:
      - Requests
  /user/profile:
    get:
      description: Получение профиля пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"